Move loginWithGoogle to parent component instead of ProviderLoginButton
